app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.0
httpx>=0.25.0
langchain>=0.0.300
langchain-google-genai>=0.0.5
langgraph>=0.0.20
sse-starlette>=1.0.0
google-adk>=0.3.0
jwcrypto>=1.5.0
asyncclick>=8.1.0 
PyJWT>=2.0.0